fix bugs

+++ database/migrations/2021_08_22_144422_create_products_table.php
@@ -27,9 +27,9 @@ class CreateProductsTable extends Migration
             $table->tinyInteger('status')->default(0);
             $table->tinyInteger('marketable')->default(1)->comment('1 => marketable, 0 => is not marketable');
             $table->string('tags');
-            $table->tinyInteger('sold_number')->default(0);
-            $table->tinyInteger('frozen_number')->default(0);
-            $table->tinyInteger('marketable_number')->default(0);
+            $table->integer('sold_number')->default(0);
+            $table->integer('frozen_number')->default(0);
+            $table->integer('marketable_number')->default(0);
             $table->foreignId('brand_id')->constrained('brands')->nullable()->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('category_id')->constrained('product_categories')->nullable()->onUpdate('cascade')->onDelete('cascade');
             $table->timestamp('published_at');
